Refactor sidebar onramp button to use custom hook
